% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{datum_available}
\alias{datum_available}
\title{Test whether datum is available}
\usage{
datum_available(executable = NULL)
}
\arguments{
\item{executable}{Optional explicit executable path.}
}
\value{
One non-missing logical value.
}
\description{
Test whether datum is available
}
\examples{
datum_available()
}
