% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datur-package.R
\docType{package}
\name{datur-package}
\alias{datur}
\alias{datur-package}
\title{datur: An R interface to datum}
\description{
\code{datur} runs the external \code{datum} data update checker without shell
interpolation and turns its JSON output into stable R objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jprybylski/datur}
  \item \url{https://jprybylski.github.io/datur/}
  \item Report bugs at \url{https://github.com/jprybylski/datur/issues}
}

}
\author{
\strong{Maintainer}: John Prybylski \email{jprybylski@gmail.com} (\href{https://orcid.org/0000-0001-5802-0539}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item John Prybylski \email{jprybylski@gmail.com} (\href{https://orcid.org/0000-0001-5802-0539}{ORCID}) [copyright holder]
}

}
\keyword{internal}
